requirements.txt
Django==5.1.3
djangorestframework==3.15.2
boto3==1.35.68
django-cors-headers==4.6.0